新增LLMGround包初始化，更新依赖项以支持huggingface-hub，重构Controller类以支持异步操作，完善worker心跳和注册功能，优化代码结构。

+++ dataModels/WorkerInfo.py
@@ -8,4 +8,4 @@ class WorkerStatus(BaseModel):
 class WorkerInfo(BaseModel):
     worker_addr: str
     worker_status: WorkerStatus
-    last_heart_beat_time: int = -1
+    last_heart_beat_time: float = -1
